Add formatting to Action! and footer
The formatting is complete, except for the button in the
call to action section. For some reason, it will not changes
its formatting. In-line formatting works, sure. But that's
not the way to do things. Out of the 10 or so properties that
I have tried to change, only the border and padding change. But
they do so weirdly. I will reboot and rest. I will take a look
at it again in a few minutes.