Style the button and header links
I found out why the styles were not applied to the button:
I was applying the style to the div that contained the button
rather than on the button itself. The header links were already
sitting in the center of their div, so I simply increased their
margin. That way, the text seems nicer.

It was a nice project to work on, and I definitely learned a
lot since creating my Odin Recipe web page. How long ago was that?
It is hard to keep track of time, but as Ferris Bueller said:
Life moves pretty fast. If you don't stop and look around once in
a while, you could miss it.